Simplify method call and string-scrub inculsion
I was in screaming hurry when I did both of these the first time, and so
did not do them well--I relied on the ruby version, rather than checking
for the existence of the method.  I also added unnecessary swtich logic
to send the method to older version of Ruby, where the prepend method is
private.  This commit resolves that technical debt. (facepalm).